+++ apps/docs/src/app/core/core-documentation.routes.ts
@@ -113,6 +113,13 @@ export const ROUTES: Routes = [
                     )
             },
             {
+                path: 'formatted-text',
+                loadChildren: () =>
+                    import('./component-docs/formatted-text/formatted-text-docs.module').then(
+                        (m) => m.FormattedTextDocsModule
+                    )
+            },
+            {
                 path: 'global-config',
                 loadChildren: () =>
                     import('./component-docs/global-config/global-config-docs.module').then(
@@ -193,6 +200,13 @@ export const ROUTES: Routes = [
                     )
             },
             {
+                path: 'message-toast',
+                loadChildren: () =>
+                    import('./component-docs/message-toast/message-toast-docs.module').then(
+                        (m) => m.MessageToastDocsModule
+                    )
+            },
+            {
                 path: 'multi-input',
                 loadChildren: () =>
                     import('./component-docs/multi-input/multi-input-docs.module').then((m) => m.MultiInputDocsModule)
@@ -202,6 +216,13 @@ export const ROUTES: Routes = [
                 loadChildren: () =>
                     import('./component-docs/notification/notification-docs.module').then(
                         (m) => m.NotificationDocsModule
+                    )
+            },
+            {
+                path: 'object-identifier',
+                loadChildren: () =>
+                    import('./component-docs/object-identifier/object-identifier-docs.module').then(
+                        (m) => m.ObjectIdentifierDocsModule
                     )
             },
             {
@@ -326,6 +347,14 @@ export const ROUTES: Routes = [
             {
                 path: 'tree',
                 loadChildren: () => import('./component-docs/tree/tree-docs.module').then((m) => m.TreeDocsModule)
+            },
+            {
+                path: 'object-number',
+                loadChildren: () => import('./component-docs/object-number/object-number-docs.module').then((m) => m.ObjectNumberDocsModule)
+            },
+            {
+                path: 'wizard',
+                loadChildren: () => import('./component-docs/wizard/wizard-docs.module').then((m) => m.WizardDocsModule)
             }
         ]
     }