+++ libs/core/src/lib/object-identifier/object-identifier.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { ObjectIdentifierComponent } from './object-identifier.component';
+
+@NgModule({
+    declarations: [ObjectIdentifierComponent],
+    imports: [CommonModule],
+    exports: [ObjectIdentifierComponent]
+})
+export class ObjectIdentifierModule {}