+++ libs/platform/src/lib/platform.module.ts
@@ -23,6 +23,7 @@ import { PlatformInputModule, PlatformStepInputModule, PlatformInputGroupModule 
 import { PlatformComboboxMobileModule } from './components/form/combobox/combobox-mobile/combobox-mobile.module';
 import { PlatformComboboxModule } from './components/form/combobox';
 import { PlatformThumbnailModule } from './components/thumbnail/public_api';
+import { PlatformObjectMarkerModule } from './components/object-marker/object-marker.module';
 
 @NgModule({
     imports: [CommonModule, FundamentalNgxCoreModule],
@@ -51,7 +52,8 @@ import { PlatformThumbnailModule } from './components/thumbnail/public_api';
         PlatformInputGroupModule,
         PlatformComboboxModule,
         PlatformComboboxMobileModule,
-        PlatformThumbnailModule
+        PlatformThumbnailModule,
+        PlatformObjectMarkerModule
     ]
 })
-export class FundamentalNgxPlatformModule { }
+export class FundamentalNgxPlatformModule {}